Fix `SOURCE_MAPPING_PATTERN` to correctly match CSS comments
Previously the `*` was not escaped in this Regexp, so it was being
interpreted as a "zero or more" quantifier on the previous character,
rather than as a literal `*` as intended. This meant that the Regexp
matched the prefix `# sourceMappingURL=` even when it did not appear
inside a CSS comment. Now it'll only match `/*# sourceMappingURL=`.