Raise MissingAssetError in compute_asset_path instead of each resolver
`compute_asset_path` is the ActionView method that asset pipeline gems
are expected to implement [1] [2], so it's important for it to raise an
exception when the asset doesn't exist in order to prevent typos from
going unnoticed when developers invoke an ActionView helper (e.g.
`asset_path`, `javascript_include_tag`, `stylesheet_link_tag`).

In 0fd78144d70c987cf33da97b4b6471ca322b021e, each resolver's `resolve`
method was updated to raise a `MissingAssetError` to handle this. But we
can also handle it by raising directly in Propshaft's
`compute_asset_path` implementation. I believe that approach is
preferable for a few reasons. The simplest reason is that it requires
less code.

Additionally, because `compute_asset_path` is the entry point for anyone
looking to understand how Propshaft integrations with the ActionView
helpers, I think it's helpful to include the exception it raises right
there in the method's definition. This is also consistent with
sprockets-rails, which raises `AssetNotFound` directly in its
`compute_asset_path` implementation [3].

But most importantly, because `compute_asset_path` is the highest level
of Propshaft's integration with Rails (i.e. the layer just before it
hands results back to Rails), I believe it's safest to raise the missing
asset error there. Consider a Propshaft extension which monkey patches
one of the internal Resolver classes, or introduces a new one, in a way
that makes it possible for `resolver.resolve` to return `nil` again. In
that case, having the `raise` in `compute_asset_path` would still
protect application developers by ensuring that the missing asset error
is still surfaced. As a corollary, authors of such an extension wouldn't
have to worry about raising `MissingAssetError` themselves since
Propshaft would handle it.

Thus, I think it makes sense to move the `raise MissingAssetError` calls
out of Propshaft's internal Resolver classes. Propshaft users aren't
expected to interact directly with instances of those internal classes,
so I think it's fine for their `resolve` methods to return `nil`, as the
`if`-expressions naturally did prior to
0fd78144d70c987cf33da97b4b6471ca322b021e.

[1]: https://github.com/rails/rails/blob/099289b6360ac82d1e0fa0a5592ac10cfc05e6e0/actionview/lib/action_view/helpers/asset_url_helper.rb#L132-L133
[2]: https://github.com/rails/rails/blob/099289b6360ac82d1e0fa0a5592ac10cfc05e6e0/actionview/lib/action_view/helpers/asset_url_helper.rb#L262-L265
[3]: https://github.com/rails/sprockets-rails/blob/118ce60b1ffeb7a85640661b014cd2ee3c4e3e56/lib/sprockets/rails/helper.rb#L84