Turn Propshaft::Server into a proper middleware
Fix: https://github.com/rails/rails/issues/55508

Mounting Propshaft::Server in the Rails router isn't suitable
because it end up executed after a long list of middlewares that
may set cookies or do other transformation we don't want.

In production, propshaft assets are served by `ActionDispatch::Static`
(or a reverse proxy), so we should try to be as close to that as possible.