Update asset digests when sourcemaps change
When sourcemaps change, their digest will change, and the generated URL in the `# sourceMappingURL` comment will change. Therefore, it's important that the digest of the original asset also changes, so that caches can be invalidated correctly.

This change makes use of the existing `#referenced_by` API in the SourceMappingUrls compiler, and implements the tests in a similar way to 59406abcfa479468f1865addc7f636a6dd91c400.